Address org switcher review feedback
